requirements.txt
pillow >= 9.3.0